requirements.txt
discord.py==2.3.2
audioop-lts==0.2.1
gspread==6.1.2
google-auth==2.29.0
python-dotenv==1.0.1
pytz==2024.1
